requirements.txt
Django==3.2.5
djangorestframework==3.12.4
whitenoise==5.3.0
Pillow==8.2.0
requests==2.25.1
tensorflow==2.8.0
numpy==1.21.0
scikit-learn==0.24.2
pandas==1.3.0
joblib==1.1.0
django-cors-headers==3.10.0
matplotlib==3.4.3
seaborn==0.11.2


# To install all the required packages with Python 3.11, run the following command:
# python3.11 -m pip install -r requirements.txt





